Added override functions

+++ Calculator/ICalculator.java
@@ -3,13 +3,13 @@
  */
 
 public interface ICalculator {
-  public int addition(int value1, int value2);
+  public double addition(double value1, double value2);
 
-  public int subtraction(int value1, int value2);
+  public double subtraction(double value1, double value2);
 
-  public int division(int value1, int value2);
+  public double division(double value1, double value2);
 
-  public int multiplication(int value1, int value2);
+  public double multiplication(double value1, double value2);
 
   public void display();
 };
